+++ node_js/index.js
@@ -2,10 +2,10 @@
 
 const parseArgs = require('minimist');
 
-const { getBuildTime, getDependencies } = require('./general_checks.js');
-const { checkInspect } = require('./quality.js');
-const { checkCoverage } = require('./coverage_nyc.js');
-const { buildMetrics, saveMetrics } = require('../git/save_metrics');
+const { getBuildTime, getDependencies } = require('./metrics/general_checks.js');
+const { checkInspect } = require('./metrics/quality.js');
+const { checkCoverage } = require('./metrics/coverage_nyc.js');
+const { buildMetrics, saveMetrics } = require('./utils/save_metrics');
 
 const BUILD_TIME = 'build-time';
 const DIRECT_DEPENDENCIES = 'direct-dependencies';
@@ -13,14 +13,15 @@ const INDIRECT_DEPENDENCIES = 'indirect-dependencies';
 const CODE_COVERAGE = 'code-coverage';
 const CODE_QUALITY = 'code-quality';
 
-const NODE_METRICS_URL = '';
-const NODE_TECH = 'node';
+const NODE_METRICS_URL = 'https://backendmetrics.engineering.wolox.com.ar/metrics';
+const NODE_TECH = 'node_js';
 const ENV_BRANCH = 'development';
 // const repository = process.env.CIRCLE_PROJECT_REPONAME || 'YOUR_REPOSITORY';
 // const branch = process.env.CIRCLE_BRANCH || 'YOUR_BRANCH';
 
 const getArgs = () => {
   const args = parseArgs(process.argv);
+  console.log(args);
   return {
     tech: args.tech || args.t || NODE_TECH,
     branch: args.branch || args.b || ENV_BRANCH,
