+++ node_js/metrics/general_checks.js
